👌 IMPROVE: Prevent name collisions by using random numbers
If different plugins both use cgb and create a block with the same name,
this would otherwise create a name collision.

Note that this is not the same as registering a block with the same name
in JS:
- PHP functions with the same name will crash in PHP, which will prevent
the plugin from even loading.
- People might change the name of the registered JS block, but might
keep the old function names in `init.php` (this already happened to us
and a different plugin author, we both initially called our block `block`).